biznisweb-mcp-server/pyproject.toml
[project]
name = "biznisweb-mcp-server"
version = "0.1.0"
description = "MCP server for BizniWeb e-shop GraphQL API"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.1.2",
    "gql[all]>=3.5.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
biznisweb-mcp = "biznisweb_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["biznisweb_mcp"]